chore: move supporting functions to separate file
Cluster was hard to read with all the supporting functions making the file super long. This way it is clearer what elements are part of the terraform config and what elements are supporting
